requirements.txt
reason_final.txt


# Web framework (if needed for API)
fastapi
uvicorn

# Data processing
pandas
python-dateutil
openpyxl

# Google ADK and GenAI (core chatbot dependencies)
google-adk
google-genai